If key is null, GetEntry will also check for items with any key.
GetEntry(service, key) behaves as it did previously, with one exception:
when key is null, it first tries to perform the old behavior (i.e. searching
for an item with the proper type and key set to null) but, if it doesn't
find anything, it'll try to find an item of the proper type with any key.