NavigationPageAdapter now takes ViewModel type into account when pushing view
PushAsync now takes the ViewModel type as an argument (defaults to
null).  If the viewmodel type is specified, it will use IoC to get an
instance of it, as opposed to relying on
ViewModelLocator.LocateForView(which ultimately just uses IoC, anyways).
No reason to look up the viewmodel type again, when we already know it.