Add FAQ for obscure SDK missing error message
Issue #648 did not state the exact error that occurred when opening the solution but if the SDK is missing (as described in https://github.com/novotnyllc/MSBuildSdkExtras/issues/190) the solution fails to load.

This is an obscure error message that is hard to search for within this repository. The modification adds a simple and prominent FAQ that explains the problem and how to fix it.